refactor(users): isolate removal cleanup workflow
